Move utility functions in the new file
